2006-08-30	Joel Sherrill <joel@OARcorp.com>
	* libmisc/cpuuse/cpuuse.c, libmisc/monitor/mon-command.c,
	libmisc/monitor/mon-driver.c, libmisc/monitor/mon-symbols.c,
	libmisc/rtmonuse/rtmonuse.c, libmisc/stackchk/check.c,
	libnetworking/libc/res_debug.c, telnetd/telnetd.c: Remove printf
	format warnings.

+++ cpukit/libnetworking/libc/res_debug.c
@@ -534,7 +534,7 @@ p_time(u_int32_t value) {
 	static char nbuf[40];
 
 	if (ns_format_ttl(value, nbuf, sizeof nbuf) < 0)
-		sprintf(nbuf, "%" PRIu32, value);
+		sprintf(nbuf, "%" PRIu32, (uint32_t) value);
 	return (nbuf);
 }
